Added remote control feature
From now there is no need to connect the device to the computer inorder to
transfer the kernel module. The android device needs to have a wifi connection.
Using the stapandroid tool  modules can be transfered to the device with
simply one command.
Besides the tool enables remote control of systemtap. So far a module can be
started, stopped or deleted.